pyproject.toml
[tool.poetry]
name = "botcommerce"
version = "0.1.0"
description = ""
authors = [ "jakhongir <ganiyevuzb@gmail.com>", "habibullo <bubbless7456@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
django-mptt = "^0.14.0"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
django-filter = "^22.1"
djangorestframework-simplejwt = "^5.2.2"
django-jazzmin = "^2.6.0"
python-dotenv = "1.0.0"
psycopg2-binary = "^2.9.5"
pillow = "^9.4.0"
django-parler = "^2.3"
sentry-sdk = "^1.14.0"
django-multiselectfield = "^0.1.12"
pyyaml = "^6.0"
faker = "^17.3.0"
pytest = "^7.2.1"
coverage = "^7.1.0"
model-bakery = "^1.10.1"
pytest-cov = "^4.0.0"
django-parler-rest = "^2.2"
celery = "^5.2.7"
django-debug-toolbar = "^3.8.1"
flake8 = "^6.0.0"
mypy = "^1.0.0"
pytest-django = "^4.5.2"
django-hosts = "^5.2"
httpx = "^0.23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
